cash-register table now display receipts instead of isolated items. reuse columns from receipts table here
